button experiments + relevant bs3 docs in comment
http://getbootstrap.com/css/#buttons

Context-specific usage
While button classes can be used on <a> and <button> elements, only
<button> elements are supported within our nav and navbar components.

Context-specific usage
Like the standard button classes, .navbar-btn can be used on <a> and
<input> elements. However, neither .navbar-btn nor the standard button
classes should be used on <a> elements within .navbar-nav.

Non-nav links
For folks using standard links that are not within the regular navbar
navigation component, use the .navbar-link class to add the proper
colors for the default and inverse navbar options.

Disabled state
Make buttons look unclickable by fading them back 50%.

If you add the disabled attribute to a <button>, Internet Explorer 9
and below will render text gray with a nasty text-shadow that we cannot
fix.

Anchor element
Add the .disabled class to <a> buttons.

This class will only change the <a>'s appearance, not its
functionality. Use custom JavaScript to disable links here.

To align components in navbars with utility classes, use .navbar-left
or .navbar-right instead.

http://getbootstrap.com/components/#navbar-component-alignment